Fix warn on save on recipe/recipebook
Previously warn on save had been fixed so that it didn't warn when
pressing the save button, however, this had only been fixed for recipes
and not for recipebooks.

Furthermore, this fix had not considered API errors which meant that the
feature would not work after such an unsuccessful save.